+++ src/App.css
@@ -1,38 +1,40 @@
-.App {
+body {
   text-align: center;
+  background: black;
+  color: whitesmoke;
+  margin: 0
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+ul {
+  display: flex;
+  flex-wrap: wrap;
+  list-style: none;
+  justify-content: center;
+  gap: 10px;
+  padding: 0;
+  margin: 0 10px;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+li {
+  width: 300px;
+  border: whitesmoke 2px solid;
+  padding: 10px;
   display: flex;
   flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+
 }
 
-.App-link {
-  color: #61dafb;
+button {
+  font-family: 'Times New Roman', Times, serif;
+  border-radius: 0px;
+  justify-self: flex-end;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+input {
+  margin: 0 5px;
+  border-radius: 0;
+}
+
+a {
+  color: whitesmoke;
 }
